requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
pip==20.2.1
requests==2.28.2
setuptools==49.2.1
urllib3==1.26.14
python-telegram-bot==20.0a2